requirements.txt
crewai
langchain
langchain-openai
openai
FlagEmbedding
pymilvus
streamlit
python-dotenv
numpy
selenium
scrappy